requirements.txt
asgiref==3.11.1
Django==6.0.3
django-filter==25.2
djangorestframework==3.17.0
mysqlclient==2.2.8
PyMySQL==1.1.2
python-decouple==3.8
sqlparse==0.5.5
